cleanup: move BC code to its own header file, create common.hpp, remove get from accessor name
