+++ src/common/boundary_condition.cpp
@@ -4,10 +4,11 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
+#include "common/boundary_condition.hpp"
+
 #include <algorithm>
 
 #include "common/logger.hpp"
-#include "common/serac_types.hpp"
 
 namespace serac {
 
@@ -70,7 +71,7 @@ void BoundaryCondition::project(mfem::ParGridFunction& gf, const mfem::ParFinite
 void BoundaryCondition::project() const
 {
   SLIC_ERROR_IF(!state_, "Boundary condition must be associated with a FiniteElementState.");
-  project(*(state_->gf), *(state_->space));
+  project(*state_->gf, *state_->space);
 }
 
 void BoundaryCondition::projectBdr(mfem::ParGridFunction& gf, const double time, const bool should_be_scalar) const
@@ -96,10 +97,10 @@ void BoundaryCondition::projectBdr(mfem::ParGridFunction& gf, const double time,
 void BoundaryCondition::projectBdr(const double time, const bool should_be_scalar) const
 {
   SLIC_ERROR_IF(!state_, "Boundary condition must be associated with a FiniteElementState.");
-  projectBdr(*(state_->gf), time, should_be_scalar);
+  projectBdr(*state_->gf, time, should_be_scalar);
 }
 
-void BoundaryCondition::eliminateFrom(mfem::HypreParMatrix& k_mat)
+void BoundaryCondition::eliminateFromMatrix(mfem::HypreParMatrix& k_mat)
 {
   SLIC_ERROR_IF(!true_dofs_, "Can only eliminate essential boundary conditions.");
   eliminated_matrix_entries_.reset(k_mat.EliminateRowsCols(*true_dofs_));
