+++ src/solvers/nonlinear_solid_operators.hpp
@@ -15,7 +15,7 @@
 
 #include <memory>
 
-#include "common/serac_types.hpp"
+#include "common/common.hpp"
 #include "mfem.hpp"
 
 namespace serac {
